Comandos por voz en libro

+++ forms/libro/form_libro.py
@@ -2,10 +2,10 @@ from tkinter import messagebox
 from db.BD import BaseDatos
 from forms.libro.form_libro_designer import FormLibroDesigner
 from tkinter import filedialog
-# import util.aruco as ar
 import util.aruco as ar
 import util.speech as speech
 from threading import Thread
+import speech_recognition as sr
 
 class FormLibro(FormLibroDesigner):
     def __init__(self, basedatos : BaseDatos, id_libro=None, id_usuario=None):
@@ -25,12 +25,14 @@ class FormLibro(FormLibroDesigner):
             messagebox.showerror(message=str(e), title="Mensaje")
     
     def guardarCategoria(self, id_usuario, id_libro, id_categoria):
-        if not (self.bd.existeEnBiblioteca(id_usuario=id_usuario, id_libro=id_libro, id_categoria=id_categoria)):
-            self.bd.insertarEnBiblioteca(id_usuario=id_usuario, id_libro=id_libro, id_categoria=id_categoria)
-            messagebox.showinfo(title="Libro añadido", message="Se ha añadido el libro a la sección correspondiente")
-        else:
-            self.bd.eliminarDeBiblioteca(id_usuario=id_usuario, id_libro=id_libro, id_categoria=id_categoria)
-            messagebox.showinfo(title="Libro eliminado", message="Se ha eliminado el libro de la sección correspondiente")
+        if not self.comando_ejecutandose:
+            self.comando_ejecutandose = True
+            if not (self.bd.existeEnBiblioteca(id_usuario=id_usuario, id_libro=id_libro, id_categoria=id_categoria)):
+                self.bd.insertarEnBiblioteca(id_usuario=id_usuario, id_libro=id_libro, id_categoria=id_categoria)
+                messagebox.showinfo(master=self.ventana, title="Libro añadido", message="Se ha añadido el libro a la sección correspondiente")
+            else:
+                self.bd.eliminarDeBiblioteca(id_usuario=id_usuario, id_libro=id_libro, id_categoria=id_categoria)
+                messagebox.showinfo(master=self.ventana, title="Libro eliminado", message="Se ha eliminado el libro de la sección correspondiente")
     
     def realidadAumentada(self, imagen):
         print(ar.camaraEjecutandose)
@@ -41,3 +43,5 @@ class FormLibro(FormLibroDesigner):
 
     def leerDescripcion(self, texto):
         speech.texto_a_audio(texto=texto)
+    
+
